revert to Int

+++ src/main/java/general/ParseInt.java
@@ -4,26 +4,26 @@ import java.util.function.Function;
 
 public class ParseInt {
 
-    public static long parseStringToInt(final String number) {
-        final Function<Character, Long> charToInt = (final Character character) ->
+    public static int parseStringToInt(final String number) {
+        final Function<Character, Integer> charToInt = (final Character character) ->
             switch (character) {
-                case '0' -> 0L;
-                case '1' -> 1L;
-                case '2' -> 2L;
-                case '3' -> 3L;
-                case '4' -> 4L;
-                case '5' -> 5L;
-                case '6' -> 6L;
-                case '7' -> 7L;
-                case '8' -> 8L;
-                default -> 9L;
+                case '0' -> 0;
+                case '1' -> 1;
+                case '2' -> 2;
+                case '3' -> 3;
+                case '4' -> 4;
+                case '5' -> 5;
+                case '6' -> 6;
+                case '7' -> 7;
+                case '8' -> 8;
+                default -> 9;
         };
 
-        long value = 0L;
-        long power = 1L;
+        int value = 0;
+        int power = 1;
         for(int i = number.length() - 1; i >= 0; i--) {
             value += (charToInt.apply(number.charAt(i)) * power);
-            power *= 10L;
+            power *= 10;
         }
 
         return value;
